test: Move DirsAndFileName functionality tests with the definition
